refactor(Labels): Move "groupBy" logic out of SceneByVariableRepeaterGrid
